+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+- `create` and `recreate` commands without an argument will create a virtual 
+  environment with the same interpreter that started `vien`, instead of 
+  the interpreter available as `python3`
+
 # 5.0
 
 - option `--project-dir` aka `-p` is now supported by all commands